Add AVXVNNI support
Intel, hopping on the ML all the things band wagon, decided that VNNI
was just _too_ good to not include on their lesser client CPUs without
AVX512. Given how many cycles it saves for the 32 bit widening
operation and how close we already were to something that supports this
sans AVX512, we may as well include support for it. To actually get
maximal ILP for this, this loop had to be unrolled a considerable
amount, but we're finally faster on both P cores and E cores for this
implementation. The gains are minor but measurable.

The copying variant is only faster on E-cores, so it's not being
compiled or used. This may need further testing on MSVC, they have no
definitive "arch" flag for it as far as I can tell.