riscv: features: Detect extensions always and only through AT_HWCAP
Previously, RVV/Zbc is used when

- Linux kernel version (detected by uname(2)) is greater or equal to
  6.5, and HWCAP claims support for such extension, or
- Linux kernel version is less than 6.5, but compiler claims support for
  the extension

and for RVV, we additionally check whether it's (the earlier, draft)
v0.7.1 release or the (the later, ratified) v1.0 release. Zlib-ng only
supports v1.0.

This strategy is problematic in several aspects,

- Fallback to compile-time constants with Linux kernel earlier than 6.5
  makes it hard for distributions to ship the same binary for different
  machines with different kernel versions.
- Detecting Linux kernel version through uname(2) could violate sandbox
  enforcement[1].

And since RVV introduces special vector registers, RVV usage in
userspace relies on the kernel to implement proper context switching,
which is only present since Linux kernel 6.5, which is also the version
when COMPAT_HWCAP_ISA_V was introduced. This makes it reasonable to only
do HWCAP-based extension detection for RVV.

The additional check for RVV version is also unnecessary since in
mainline Linux kernel, COMPAT_HWCAP_ISA_V is only set when a ratified
v1.0 implementation of RVV is present. A downstream kernel claiming
COMPAT_HWCAP_ISA_V with a v0.7.1 RVV implementation breaks the kABI,
and userspace shouldn't pay for the mistake.

This patch removes extra detection logic, now RVV is and only is used
when the kernel claims COMPAT_HWCAP_ISA_V. The detection logic of Zbc
checks an undefined HWCAP bit, thus doesn't detect Zbc correctly with
any current Linux kernel, which should be fixed with another patch.

Fixes: 6ff8b52cefe5 ("Support RVV hwcap detect at runtime")
Link: https://github.com/openssh/openssh-portable/pull/605 # [1]
Closes: https://github.com/zlib-ng/zlib-ng/issues/1705
Closes: https://github.com/zlib-ng/zlib-ng/issues/1769
Signed-off-by: Yao Zi <ziyao@disroot.org>